requirements.txt
flask==3.0.3
gradio==4.44.0
requests==2.32.3
httpx==0.27.2
python-dotenv==1.0.1
sqlalchemy==2.0.34
alembic==1.13.2
pydantic==2.9.2
nltk==3.9.1
networkx==3.3
radon==6.0.1
uvicorn==0.30.6
gunicorn==22.0.0
pytest==8.3.3
pytest-cov==5.0.0
black==24.8.0
ruff==0.6.8
transformers==4.44.2
peft==0.11.1
accelerate==0.34.2
datasets==2.21.0
evaluate==0.4.2
sacrebleu==2.4.3
rouge-score==0.1.2
torch==2.4.0




